Revert "controls: Fix the ListItem height"
This reverts commit 1f9744fb3a2b208d7376e65226d07e4ea44312db.

The aforementioned commit makes list item shorter when they have
a leftItem and/or a rightItem that is taller than sub text.

This behavior can be observed with the settings power module
where some list items have a ComboBox in the rightItem which
is obviously taller than the text but the overall list item
height doesn't include the ComboBox.

Closes: #32