core: Fix invalid URI schema returned by Utils.getSourceForIconName()
The logic behind this function is to return a icon:// URI that will
be replaced by a full path to the Fluid material icons by Icon,
but this is wrong since applications using this method will never
replace the URI with a relative path to the icons and fail because
icon:// is an invalid schema.

Instead we implement an image provider that will load the requested
icon be it from a system-widge installation or resources.

Closes: #45