controls: Fix issues introduced with TabbedPage
Current implementation can be improved moving TabBar to the TabbedPage
itself. As a matter of fact, Page has a header property that we can use
to place the TabBar instead of hacking the AppBar to contain two components.

We also reduce layout complexity and go back to anchoring AppBar contents
to the parent.

Height is now calculated correctly.

Issue: #55