Refactor source tree and add qmake build system
Move source code around in a similar fashion of Qt modules and
add the qmake build system.

When building with cmake, install headers in the same location as qmake
in order to make both coexist.

Replace all foreach with C++11 for loops to avoid build failures.

Recreate plugins.qmltypes with dependencies to avoid repeating declarations.

Add MD icons as a git submodule.

Build documentation with CI.

[ChangeLog][Build System] Add qmake build system.