Initial QBS support
This build system is hopefully the future of Qt, given that qmake is
hard to use. Another benefit is qbs is not the usual makefile generator.

That being said, the effort done until now allows us to make dynamic builds
from QtCreator and command line but it lacks:

* qmake .pri generation
* cmake module generation
* pkg-config generation
* autotests do not run (missing *.qml files)
* installation (copy files from install-root to a path)

After having implemented the above features we also need to support
per-project installations and static builds of fluid-demo.

This effort serves to learn how qbs works, the know-how acquired here
will be useful for all the other Liri projects that usually just
need the classic system-wide installation, so that is our priority.