Add icons to the project
Running the script is problematic on Windows for the lack of bash.
Another issue is fetch such a large repository as the material-design-icons
one, so we are better off we importing the icons by launching the script
from time to time.

The whole icons/ directory is just a few KiB and a manual update can be
done at release time (it would have been needed with the git submodule
anyway).

Closes: #95
Closes: #96