TabBar of TabbedPage is not flicking
Even when the width of the tabs exceed the width of page,
the TabBar does not flick.

To fix it either an explicit width could be specified
to the TabBar or the TabBar could be made to occupy
the full width by specifying the right anchor,
(as left anchor was already specified). The latter
option was taken.

Closes: #118