Fix wave not correctly hiding UI elements
Previously a Wave would only cut UI elements in a rectangular shape.
This problem is now fixed and UI elements are correctly cut in a
circular shape.

Another problem of the old Wave was, that it would change the position
of inner components depending on its state, which caused the programmer
to manually place components inside a Wave depending on the wave's size.

This also was a performance problem, since it created unnecessary
bindings for the x and y properties.

Those problems are all fixed now, by fundamentally changing the
underlying way, that a Wave cuts its children.

Previously it would just use a Rectangle with a radius property and clip
set to true. Now it uses a real OpacityMask, which basically is just a
small OpenGL shader.

Except for the improvement, that the user of a Wave isn't forced to
adapt the position of inner components depending on the wave's size
anymore, the exposed object hasn't changed, so a Wave can be used just
the same way as before.