Refactor ListItem
Merge BaseListItem and ListItemDelegate into ListItem in order to
use the QtQuick Controls 2 ripple effect.

This affects also Subheader which used BaseListItem to make non
interactable which is now achieved simply replacing the background
item with an empty Item with the appropriate implicitHeight.

This change might also help with issue #102. Previously our Ripple
was used but it inherits from MouseArea and so it eats mouse
input which might cause the issue.