Fix Wave being in an invalid intial state
Previously a Wave was visible in its initial state regardless of whether
it was open or not. The wave had to be opened and closed again to become
not visible.

Now the Wave is immediately invisible, if the initial state is unopened.
Therefore components behind the Wave receive GUI events, even if the
Wave is still in its initial state.

Also the open property is now explicitly false to improve readability.