SearchBar: Do not unnecessarily catch input events
The SearchBar is always visible and unnecessarily used a Flickable
control which caught input events. Now the SearchBar uses an Item
instead and does not steal input events.

Also the SearchBar accessed an item with id searchResults although it
was not defined in the same file, which caused an user unfriendly
dependency on an item with this exact id which had to be added by the
user.
Fixed by creating a property for this purpose.