Use system qbs-shared by default
When building all Liri projects together or when qbs-shared is installed
on the system, qbs finds two copies of those modules and since v1.10
will complain.

With this property we don't use the submodule copy by default and
expect a "superbuild" (all Liri projects together) or qbs-shared
installed on the system.