SearchBar: Added new property and improved keyboard handling
Add expanded property to inform users whether the search bar
is expanded or not.

While we are here we improve keyboard handling.

The spell suggestions/predictive text feature  in the
SearchBar's TextInput prevents the trigger of onTextChanged
unless the user accepts a spell suggestion for presses space
bar. turning off the predictive text feature seems to
improve the usability in mobiles.

The continued focus on the TextInput of SearchBar causes the
onscreen keyboard to open up even if the SearchBar is closed.
This problem is pronounced in mobiles. We have released the
focus on the TextInput when the SearchBar is closed.