Fluid.Controls plugin doesn't really depend on Fluid.Core
The dependency was here to force a build order which is something that
qbs should figure out itself.

A dependency like this is also problematic because it will link the
Fluid.Controls plugin to the one of Fluid.Core.
I could set cpp.link: false to avoid that but still the dependency
doesn't make much sense.