Add path to Fluid build to per-project qmake example
When running the example app from QtCreator the Fluid import path
is not found.

Add the path of the build that is relative to the executable,
this should fix running the app from the IDE during development.

At runtime the Fluid imports should be deployed next to the app
binary in a tree like this:

.
├── src
├── qml/
│   ├── Fluid/