TimeSelector template
Implement the TimeSelector API with C++ which is much cleaner, for
instance we can define an enum for the time selector mode and time
mode instead of using strings.

This commit introduce Fluid.Templates a new QML plugin with
templates only that is used by Fluid.Controls.

While we are here we rename a few properties to improve the API
and convert the imperative code in TimeSelector to declarative.

Issue: #165