Redesign Chip
Make Chip look more like the Material Design specifications:

Inherit from AbstractButton so we have mouse input management,
and other Control properties.

Replace the delete ToolButton with an Icon: it doesn't seem to be
a tool button from the Material Design specifications and also
change the icon to what we see in the specifications.

Change foreground and background on mouse hover.

Remove the caption property and use text instead.

Add the ability to open a popup with alternative options for
expandable chips (for example contact chips).