Expose AppBar overflow menu open state
It is useful to know if overflow menu of AppBar is open or not. For example
NavigationDrawer dragMargin can be set to 0 when menu is open to disable
dragging drawer open at the same time as menu. Depending on application
there can be other situations where overflow menu state is valuable
information.

[ChangeLog][Controls] Add property AppBar::overflowMenuVisible to know
whether the overflow menu is open or not.