Explicitely depends on Wayland::Client on Linux
On Linux we want to build a protocol client to change decoration
colors according to the Material Design specifications.

Turns out we need to explicitely depend on Wayland::Client because it
is not pulled in by Qt5::WaylandClient.

It used to works for Linux distributions that had wayland-client.h
under /usr/include (a path that is traditionally always included),
but doesn't work on OpenSuSE where it is installed under a different
path such as /usr/include/wayland.

Closes: #295